error fixed SEE DESCRIPTION
to fix the error I created a singleton object called "singleton" in RhythMixApp.kt defined as follows
Singleton:
     controller: NavHostController?
     vm: RhythmixViewModel

Whenever you are making a composable function do not pass a vm around. instead import the singleton and do singleton.vm. When doing a navigation do singleton.controller!! (not null assertion is okay since it is initialized at the very beginning and never modified after that.)